Update UMLFactory to work with new SBOLFactory architecture. Also refactor it back into the uml_factory module, replacing the nonfunctional OPIL UMLFactory that was there. Add a stub test of graphics generation Note that the graphic being generated is not yet correct.
