+++ sbol_factory/__init__.py
@@ -1,3 +1,3 @@
-from .sbol_factory import SBOLFactory, Document, ValidationReport, UMLFactory
+from .sbol_factory import SBOLFactory, Document, ValidationReport
+from .uml_factory import UMLFactory
 from .shacl_validator import ShaclValidator
-import os
