+++ sbol_factory/uml_factory.py
@@ -1,40 +1,49 @@
-from .opil_factory import OPILFactory
 from .query import Query
+
 import sbol3 as sbol
+
+from math import inf
+import posixpath
+import os
 import graphviz
 
-import rdflib
-import os
-import posixpath
-import argparse
-#import logging
-from math import inf
+class UMLFactory:
+    """
+    Class for generating UML diagrams from an ontology file
+    """
 
-class UMLFactory():
+    def __init__(self, ontology_path, ontology_namespace):
+        self.namespace = ontology_namespace
+        self.query = Query(ontology_path)
 
-    def __init__(self, opil_factory, output_path):
-        self.query = opil_factory.query
-        self.opil_factory = opil_factory
+    def generate(self, output_path):
+        if not os.path.exists(output_path):
+            os.mkdir(output_path)
         for class_uri in self.query.query_classes():
+            if self.namespace not in class_uri:
+                continue
             class_name = sbol.utils.parse_class_name(class_uri)
             dot = graphviz.Digraph(class_name)
             # dot.graph_attr['splines'] = 'ortho'
-            self.generate(class_uri, self.draw_class_definition, dot)
-            self.generate(class_uri, self.draw_abstraction_hierarchy, dot)
+
+            # Order matters here, as the label for an entity
+            # will depend on the last rendering method called
+            self._generate(class_uri, self.draw_abstraction_hierarchy, dot)
+            self._generate(class_uri, self.draw_class_definition, dot)
             source = graphviz.Source(dot.source.replace('\\\\', '\\'))
             outfile = f'{class_name}_abstraction_hierarchy'
             source.render(posixpath.join(output_path, outfile))
 
-    def generate(self, class_uri, drawing_method_callback, dot_graph=None):
-        if self.opil_factory.namespace not in class_uri:
+    def _generate(self, class_uri, drawing_method_callback, dot_graph=None):
+        if self.namespace not in class_uri:
             return ''
         superclass_uri = self.query.query_superclass(class_uri)
-        self.generate(superclass_uri, drawing_method_callback, dot_graph)
+        self._generate(superclass_uri, drawing_method_callback, dot_graph)
 
         class_name = sbol.utils.parse_class_name(class_uri)
 
-        if class_name in globals().keys():
-            return ''
+        # if class_name in globals().keys():
+        #     return ''
 
         drawing_method_callback(class_uri, superclass_uri, dot_graph)
 
@@ -42,7 +51,7 @@ class UMLFactory():
 
         subclass_uris = self.query.query_subclasses(class_uri)
         if len(subclass_uris) <= 1:
-            return 
+            return
 
         class_name = sbol.utils.parse_class_name(class_uri)
         if dot_graph:
@@ -57,15 +66,13 @@ class UMLFactory():
 
         for uri in subclass_uris:
             subclass_name = sbol.utils.parse_class_name(uri)
-            create_inheritance(dot, class_uri, uri)
-            #label = f'{subclass_name}|'
-            #label = '{' + label + '}'  # graphviz syntax for record-style label
+            #create_inheritance(dot, class_uri, uri)
             label = self.label_properties(uri)
             create_uml_record(dot, uri, label)
-
-        if not dot_graph:
-            source = graphviz.Source(dot.source.replace('\\\\', '\\'))
-            source.render(f'./uml/{class_name}_abstraction_hierarchy')
+            self.draw_class_definition(uri, class_uri, dot)
+        # if not dot_graph:
+        #     source = graphviz.Source(dot.source.replace('\\\\', '\\'))
+        #     source.render(f'./uml/{class_name}_abstraction_hierarchy')
         return
 
     def label_properties(self, class_uri):
@@ -79,7 +86,7 @@ class UMLFactory():
         associative_properties = [uri for uri in property_uris if uri not in
                                     compositional_properties]
 
-        # Initialize associative properties
+        # Label associative properties
         for property_uri in associative_properties:
             if len(associative_properties) != len(set(associative_properties)):
                 print(f'{property_uri} is found more than once')
@@ -88,10 +95,10 @@ class UMLFactory():
             lower_bound, upper_bound = self.query.query_cardinality(property_uri, class_uri)
             if upper_bound == inf:
                 upper_bound = '*'
-            object_class_uri = self.query.query_range(property_uri)
+            object_class_uri = self.query.query_property_datatype(property_uri, class_uri)
             arrow_label = f'{property_name} [{lower_bound}..{upper_bound}]'
 
-        # Initialize compositional properties
+        # Label compositional properties
         for property_uri in compositional_properties:
             if len(compositional_properties) != len(set(compositional_properties)):
                 print(f'{property_uri} is found more than once')
@@ -101,10 +108,10 @@ class UMLFactory():
             lower_bound, upper_bound = self.query.query_cardinality(property_uri, class_uri)
             if upper_bound == inf:
                 upper_bound = '*'
-            object_class_uri = self.query.query_range(property_uri)
+            object_class_uri = self.query.query_property_datatype(property_uri, class_uri)
             arrow_label = f'{property_name} [{lower_bound}..{upper_bound}]'
 
-        # Initialize datatype properties
+        # Label datatype properties
         property_uris = self.query.query_datatype_properties(class_uri)
         for property_uri in property_uris:
             property_name = self.query.query_label(property_uri).replace(' ', '_')
@@ -201,9 +208,9 @@ class UMLFactory():
             label += f'{property_name} [{lower_bound}..{upper_bound}]: {datatype}\\l'
         label = '{' + label + '}'  # graphviz syntax for record-style label
         create_uml_record(dot, class_uri, label)
-        if not dot_graph:
-            source = graphviz.Source(dot.source.replace('\\\\', '\\'))
-            source.render(f'./uml/{CLASS_NAME}')
+        # if not dot_graph:
+        #     source = graphviz.Source(dot.source.replace('\\\\', '\\'))
+        #     source.render(f'./uml/{CLASS_NAME}')
         return log
 
 
@@ -238,7 +245,7 @@ def create_association(dot_graph, subject_uri, object_uri, label):
             'fontname' : 'Bitstream Vera Sans',
             'fontsize' : '8',
             'dir' : 'both'
-        } 
+        }
     association_relationship['label'] = label
     dot_graph.edge(subject_class, object_class, **association_relationship)
     qname = format_qname(object_uri)
@@ -255,7 +262,7 @@ def create_composition(dot_graph, subject_uri, object_uri, label):
             'fontname' : 'Bitstream Vera Sans',
             'fontsize' : '8',
             'dir' : 'both'
-        } 
+        }
     composition_relationship['label'] = label
     dot_graph.edge(subject_class, object_class, **composition_relationship)
     qname = format_qname(object_uri)
@@ -271,8 +278,9 @@ def create_inheritance(dot_graph, superclass_uri, subclass_uri):
             'fontname' : 'Bitstream Vera Sans',
             'fontsize' : '8',
             'dir' : 'back'
-        } 
+        }
     dot_graph.edge(superclass, subclass, **inheritance_relationship)
     qname = format_qname(superclass_uri)
     label = '{' + qname + '|}'
-    OPILFactory.create_uml_record(dot_graph, superclass_uri, label)
+    create_uml_record(dot_graph, superclass_uri, label)
+
